Initial commit of the website repo.
This commit includes all the text files of the Davidsonian archives processed through tesseract. It also includes the application .py file, styles .css file, and .html file that compile into the website. Other files include assets of images that go on the website, packages required to run the website, etc.